fix: вынесены функции отвечающие за валидацию данных для создания класса Film
